refactor: simplify idp-basic relational tests
