+++ plugins/idp-basic/src/test/java/org/apache/gravitino/idp/basic/storage/relational/mapper/TestIdpUserMetaMapperPostgreSQL.java
@@ -59,18 +59,8 @@ public class TestIdpUserMetaMapperPostgreSQL extends IdpMapperTestBase
   }
 
   @TestTemplate
-  public void testUpdateIdpUserPasswordReturnsZeroForDeletedUser() {
-    IdpUserMetaMapperTest.super.testUpdateIdpUserPasswordReturnsZeroForDeletedUser();
-  }
-
-  @TestTemplate
   public void testSoftDeleteIdpUser() {
     IdpUserMetaMapperTest.super.testSoftDeleteIdpUser();
-  }
-
-  @TestTemplate
-  public void testSoftDeleteIdpUserReturnsZeroForDeletedUser() {
-    IdpUserMetaMapperTest.super.testSoftDeleteIdpUserReturnsZeroForDeletedUser();
   }
 
   @TestTemplate
